requirements.txt
#Core ML/DL
torch

#Use the appropriate unsloth version for your CUDA version
#For CUDA 12.1:
unsloth[cu121] @ git+https://github.com/unslothai/unsloth.git

#For CUDA 11.8:
unsloth[cu118] @ git+https://github.com/unslothai/unsloth.git

#Hugging Face ecosystem
transformers==4.41.2
datasets==2.19.2
peft==0.11.1
accelerate==0.30.1
bitsandbytes==0.43.1
trl==0.8.6

#Experiment tracking and configuration
hydra-core==1.3.2
tensorboard==2.16.2 # Offline experiment tracking

#Graph processing and visualization
networkx==3.3
pyvis==0.3.2

#Web UI for inference
gradio==4.31.5

#For evaluation
evaluate==0.4.2
rouge_score==0.1.2
sacrebleu==2.4.2

#Langchain for prompt templating
langchain==0.2.1
langchain_core==0.2.3